Move spec config models to native/v1alpha
